requirements.txt
flask==3.1.0
tensorflow
numpy
Pillow
gunicorn